+++ .storybook/stories/TableFiltering/ExampleLarge.tsx
@@ -9,7 +9,7 @@ import { SortBySelect } from './components/SortBySelect';
 import { DashboardTable } from './components/DashboardTable';
 import { FilterModal } from './components/FilterModal';
 import { FilterSearchInput } from './components/FilterSearchInput';
-import { GetDataFilters, GetDataPagination, GetDataSort } from './lib/data';
+import { GetDataFilters, GetDataPagination, GetDataSort } from './lib/getData';
 import { BusinessForAudit } from './lib/generateBusinessData';
 import { Prose } from '@ag.ds-next/react/prose';
