+++ .storybook/stories/TableFiltering/components/FilterSearchInput.tsx
@@ -1,5 +1,5 @@
 import { SearchInput } from '@ag.ds-next/react/search-input';
-import { GetDataFilters } from '../lib/data';
+import { GetDataFilters } from '../lib/getData';
 
 export const FilterSearchInput = ({
 	filters,
